debugging mth5 creation in cas04 tests
